requirements.txt
# Framework web
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# Supabase
supabase-py==0.0.3
postgrest-py==0.10.8
gotrue==1.0.2

# Variables de entorno
python-dotenv==1.0.0

# Utilidades
requests==2.31.0
urllib3==2.0.4

#pip install -r requirements.txt